move NotificationOptions to it's own file
