+++ src/background/browser-adapters/chrome-adapter.ts
@@ -1,9 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { ClientChromeAdapter } from '../../common/client-browser-adapter';
-import { BrowserAdapter, NotificationOptions } from './browser-adapter';
+import { BrowserAdapter } from './browser-adapter';
 import { CommandsAdapter } from './commands-adapter';
 import { InjectorAdapter } from './injector-adapter';
+import { NotificationOptions } from './notification-options';
 import { StorageAdapter } from './storage-adapter';
 
 export class ChromeAdapter extends ClientChromeAdapter implements BrowserAdapter, StorageAdapter, CommandsAdapter, InjectorAdapter {
