+++ src/background/browser-adapters/notification-options.ts
@@ -1,0 +1,6 @@
+export interface NotificationOptions {
+    message: string;
+    title: string;
+    iconUrl: string;
+    notificationType?: string;
+}
